Move the token tests into the suite CI actually runs
studio-backend-ci.yml runs studio/backend/tests only; nothing runs
studio/backend/hub/tests, so the four spawn_worker and retry-ladder tests
would never have executed in CI. They need no hub-only fixtures, so move
them next to the route tests they belong with.
