+++ studio/backend/hub/tests/test_download_lifecycle.py
@@ -700,105 +700,3 @@ def test_a_stall_verdict_racing_a_completed_worker_is_not_recorded(monkeypatch, 
         )
         == []
     ), "a completed worker was charged a stall it raced"
-
-
-def _spawn_env(monkeypatch, hf_token, **kwargs):
-    """Run the real spawn_worker against a fake Popen and return the child's environment."""
-    captured = {}
-
-    class _Fake:
-        pass
-
-    def _fake_popen(*_args, **popen_kwargs):
-        captured.update(popen_kwargs["env"])
-        return _Fake()
-
-    monkeypatch.setattr(download_lifecycle.subprocess, "Popen", _fake_popen)
-    download_lifecycle.spawn_worker(
-        ["--repo-id", "attacker/private-model"],
-        hf_token,
-        use_xet = False,
-        **kwargs,
-    )
-    return captured
-
-
-def test_an_api_caller_does_not_borrow_the_backend_hf_token(monkeypatch):
-    """A caller the route marked as not allowed the ambient token gets an anonymous worker, even
-    though the backend process has an HF_TOKEN of its own."""
-    monkeypatch.setenv("HF_TOKEN", "operator-secret-token")
-
-    env = _spawn_env(monkeypatch, None, allow_ambient_token = False)
-
-    assert "HF_TOKEN" not in env
-    assert env["HF_HUB_DISABLE_IMPLICIT_TOKEN"] == "1"
-
-
-def test_the_ui_still_falls_back_to_the_backend_hf_token(monkeypatch):
-    """The other half: a UI session keeps the fallback, so a private repo stays downloadable for
-    an install whose token lives in the environment rather than in Settings."""
-    monkeypatch.setenv("HF_TOKEN", "operator-secret-token")
-
-    env = _spawn_env(monkeypatch, None, allow_ambient_token = True)
-
-    assert env["HF_TOKEN"] == "operator-secret-token"
-    assert env["HF_HUB_DISABLE_IMPLICIT_TOKEN"] == "0"
-
-
-def test_an_explicit_request_token_wins_over_the_backend_one(monkeypatch):
-    monkeypatch.setenv("HF_TOKEN", "operator-secret-token")
-
-    env = _spawn_env(monkeypatch, "request-token", allow_ambient_token = True)
-
-    assert env["HF_TOKEN"] == "request-token"
-
-
-def test_an_anonymous_job_stays_anonymous_on_the_http_retry(monkeypatch, tmp_path):
-    """The recovery ladder must carry the token policy: a job started without the ambient token
-    must not pick it up when the Xet worker fails and the HTTP one takes over."""
-    monkeypatch.setattr(state_dir, "cache_root", lambda: tmp_path / "state")
-    monkeypatch.setattr(download_lifecycle.threading, "Thread", _ImmediateThread)
-    monkeypatch.setattr(download_lifecycle, "_start_stall_watchdog", lambda *a, **k: None)
-    register_worker = download_lifecycle.register_worker
-
-    registry = download_registry.DownloadRegistry()
-    key = download_registry.normalize_job_key("Org/Model")
-    assert registry.claim(
-        key,
-        download_registry.TRANSPORT_XET,
-        repo_type = "model",
-        repo_id = "Org/Model",
-        variant = None,
-        blob_hashes = frozenset({"blob"}),
-    )[0]
-    retried = []
-
-    def fake_spawn(
-        _args,
-        _token,
-        *,
-        use_xet,
-        allow_ambient_token = True,
-        **_kwargs,
-    ):
-        retried.append(allow_ambient_token)
-        return _Proc(0)
-
-    monkeypatch.setattr(download_lifecycle, "spawn_worker", fake_spawn)
-    monkeypatch.setattr(download_lifecycle, "register_worker", lambda *a, **k: True)
-    assert register_worker(
-        registry,
-        key,
-        _Proc(1, b"xet failed"),
-        hf_token = None,
-        label = "Org/Model",
-        log_prefix = "Download",
-        logger = logging.getLogger("test"),
-        repo_type = "model",
-        repo_id = "Org/Model",
-        transport = download_registry.TRANSPORT_XET,
-        watch_name = "model-watch",
-        allow_ambient_token = False,
-    )
-
-    assert retried == [False], "the HTTP retry regained the backend's token"
